+++ release-notes.md
@@ -7,7 +7,7 @@
 * 169881 - Impact Assessment (central route)
 * 177796 - Build: Finance plan agreed date - change to use FSS field
 * 147093 - Build: Due diligence checks (central route)
-
+* 161362 - Build: Project development grant (PDG) page (Central route)
 
 ---
 ## 0.11.0